use rule based predicate push down
